feat(#36): agregar funcionalidad para añadir empleados al negocio

+++ src/main/java/ispp_g2/gastrostock/Negocios/NegocioRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.Optional;
 
-public interface NegocioRepository extends CrudRepository<Negocio, Long> {
+public interface NegocioRepository extends CrudRepository<Negocio, Integer> {
 
     Optional<Negocio> findByCifNif(String cifNif);
 }
